Redesign brain-gcd game
The game module generated data for all rounds
and a common block were transferred to the engine,
the engine was launched from the game module. This has been changed.
In the game module, data is generated for one round and is not accumulated;
the engine does not start.
The necessary constants are moved outside the function.